+++ app/models/engine/expressions/expression.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Engine
+  module Expressions
+    class Expression
+      def initialize
+        # Abstract class
+      end
+    end
+  end
+end