+++ app/models/engine/expressions/grouping.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Engine
+  module Expressions
+    class Grouping < Expression
+      attr_accessor :expression
+
+      def initialize(expression)
+        @expression = expression
+      end
+    end
+  end
+end