+++ app/models/engine/expressions/literal.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+module Engine
+  module Expressions
+    class Literal < Expression
+      attr_accessor :value
+
+      def initialize(value)
+        @value = value
+      end
+    end
+  end
+end